Further core data helper splits
